Refactor Paddle, Ball, and Screen logic; clean up unused code
- Fix include path for `ComputeCenterYofRectangle`.
- Modify mouse position updates in `UIButton` for better collision handling.
- Adjust screen initialization to use `InitLogoScreen` instead of `InitGameScreen`.
- Improve `ScreenManager` switch-case formatting for clarity.
- Comment out or remove unused code in `Main_Game` and `Ball` modules.
- Refactor paddle and ball initialization, grouping parameters for improved readability.
- Change default FPS from 60 to 120 for smoother gameplay.
- Fix formatting and minor typos throughout the codebase.

+++ src/Screens/screen_menu.c
@@ -4,7 +4,7 @@
  *  Created by Squid on 10/27/2025.
 ****************************************************************/
 
-#include <io.h>
+//#include <io.h>
 
 #include "screens.h"
 #include "../../include/UI/UIButton/button.h"
@@ -44,13 +44,13 @@ void InitMenuScreen(void) {
     nextScreen = SCREEN_MENU;
 
     // Start Button
-    InitUiButton(&startButton, strButtonX,
-                 strButtonY, strButtonWidth,
-                 strButtonHeight, BLUE, DARKBLUE,
+    InitUiButton(&startButton, (float)strButtonX,
+                 (float)strButtonY, (float)strButtonWidth,
+                 (float)strButtonHeight, BLUE, DARKBLUE,
                  SKYBLUE, BLACK);
     //Quit Button
-    InitUiButton(&quitButton, quitButtonX, quitButtonY,
-        quitButtonWidth, quitButtonHeight,
+    InitUiButton(&quitButton, (float)quitButtonX, (float)quitButtonY,
+        (float)quitButtonWidth, (float)quitButtonHeight,
         RED, DARKPURPLE, SKYBLUE, BLACK);
 }
